Log key event.code instead of event.key
key contains the 'result', for example A or a depending on whether shift is held. code contains the string representation of the actual key, for example KeyA.
There is also a keycode field, which contains the numeric representation of the code, but this is filled with 229 when entering Chinese text.